requirements.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Django==4.0.4
django-bootstrap-v5==1.0.11
django-braces==1.15.0
django-embed-video==1.4.4
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
factory-boy==3.2.1
Faker==13.11.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
kw==0.0.2
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
parameterized==0.8.1
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
questions==0.7.0a4
requests==2.27.1
selenium==4.1.5
simpleeval==0.9.12
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
watchdog==2.1.7
wsproto==1.1.0
